20260314 @Mookse - rename llm.mjs
